Add HTMLProofer tests;
This patch introduces html-proofer tests to the microsite. This is a
Ruby library commonly used for testing Jekyll static sites. It's
recommended for CI [in the Jekyll documentation itself](http://jekyllrb.com/docs/continuous-integration/#the-test-script).
After `npm install` is run, the `postinstall` hook will install the gem
globally on the machine. The `npm run build` script has been updated to
run `htmlproofer` with external link checking on whatever platform is
being deployed to. The build can fail for a number of reasons related to
links in the html files. For more information on `htmlproofer`, [check
out the Github repo](https://github.com/gjtorikian/html-proofer).